Save a final version of topmed talk figures script
Rather than updating the topmed-talk-figures.R script for making figures
for the paper, I plan to create a new script for paper tables and figures.
So replace topmed-talk-figures.R with a version only slightly updated from
the original, in order to read the updated data and save files to this
directory.