Filter out deprecated traits in PIE export
The previous versions of this script had failed to filter out the traits
tagged from deprecated source study versions, which included the three
CARDIA duplicates that have been so problematic. Alter the script to
filter them out here. Update the path to the tagged variable data to
use the new version of the exported data.