Finish funded/nonfunded summary split
Fix a mistake where the funded and nonfunded tag_summary_by_review_status
tables both printed out the overall summary instead. Just update the
object name in the write.table() calls.

Also add funded and nonfunded totals to the number-of-variables section,
and add parentheses around all of the totals so that they are printed
out for viewing.