Move data cleaning code to its own script
Move code for cleaning and formatting tagging data to the new
clean-tagging-data.R script. This script now produces a final cleaned
data file that is now read into a data frame at the beginning of
paper-figures-and-tables.R for use in making figures and tables.
Modify paper-figures-and-tables.R in order to get it to run properly
with the new data source.